Finish implementation and add tests
