+++ test/databind-with-unknown-property/databind-with-unknown-property.html
@@ -4,6 +4,7 @@
     <div id="{{declared}}"></div>
     <div id="{{writtenTo}}"></div>
     <div id="{{method(writtenTo)}}"></div>
+    <div title="{{title}}"></div>
 
     <!-- Bad -->
     <div id="[[onlyReadFrom]]"></div>
